backend/chat-service/requirements.txt
Django==5.1.2
channels==4.0.0
asgiref==3.8.1
channels_redis==4.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.1
uvicorn[standard]
django-extensions==3.2.3
Werkzeug==2.3.7
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
pytest-asyncio==0.21.0
coverage>=7.5
pytest-cov==6.0.0
djangorestframework==3.15.2
daphne==4.0.0
djangorestframework-simplejwt==5.2.2
requests>=2.26.0
